Add files via upload
The running time would be O(n2)  because the function will execute i**i depending on n. Magic Square creating is the most operations intensive section of the algorithm. We need to check all entries at least once to see if the square is a magic square, any algorithm must have a running time of O(N2).

For a square to be a magic square, all rows, columns, and both diagonals must sum to the same number. There are N rows, N columns, and 2 diagonals, each of which has N entries, so you can check this property in O((2N+2)N)=O(N2) time.